refactor(daemon): drop the k8s driver's compat re-exports
Importers now reach the sandbox identity constants and the launch types
through their real owners, so driver.ts stops acting as a barrel.
