+++ packages/daemon/test/cli-reconcile.test.ts
@@ -6,7 +6,7 @@ import { afterEach, describe, expect, it } from 'vitest'
 import { K8sHttp } from '@agentconnect.md/k8s-client'
 import { closeFakeApiServers, fakeApiServer } from '@agentconnect.md/k8s-client/testing'
 import { buildEnvelope, decodeEnvelope, encode, type AnyFrame } from '@agentconnect.md/protocol'
-import { AC_LABEL_AGENT, AC_LABEL_ORG } from '../src/k8s/driver.js'
+import { AC_LABEL_AGENT, AC_LABEL_ORG } from '../src/k8s/sandbox-identity.js'
 import { SandboxApi, type SandboxClaim } from '../src/k8s/sandbox-api.js'
 import { ORPHAN_DELETE_ENV } from '../src/k8s/orphan-reconciler.js'
 import { STORE_ORPHAN_DELETE_ENV } from '../src/store/retention.js'
