+++ packages/daemon/test/direct-connect-credentials.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it, vi } from 'vitest'
 import { HOOK_REPORT_REASON_PROVIDER_AUTH_REQUIRED } from '@agentconnect.md/protocol'
 import { turnFailureCode } from '../src/acp/acp-host.js'
-import { K8sDriver, RUNTIME_GRANTS } from '../src/k8s/driver.js'
+import { K8sDriver } from '../src/k8s/driver.js'
+import { RUNTIME_GRANTS } from '../src/k8s/sandbox-identity.js'
 import { K8sApiError } from '@agentconnect.md/k8s-client'
 import type { Sandbox, SandboxClaim } from '../src/k8s/sandbox-api.js'
 import { fakeGenerations } from './fake-generations.js'
