+++ packages/daemon/test/k8s-driver.test.ts
@@ -3,7 +3,8 @@ import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { describe, expect, it, vi } from 'vitest'
 import { FakeClock } from '@agentconnect.md/connection'
-import { K8sDriver, AC_LABEL_AGENT, AC_LABEL_ORG } from '../src/k8s/driver.js'
+import { K8sDriver } from '../src/k8s/driver.js'
+import { AC_LABEL_AGENT, AC_LABEL_ORG } from '../src/k8s/sandbox-identity.js'
 import { LocalStore } from '../src/store/local-store.js'
 import { fakeGenerations } from './fake-generations.js'
 import { GuardedResumeRejectedError, OperatingModeRejectedError } from '../src/k8s/sandbox-api.js'
